requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
rapidfuzz==3.9.7
dateparser==1.2.0
